part 2 not fully working yet
you're not actually finding if something is inside and outside, if two
pipes lay right next to each other then open it's not "inside" it's just
being blocked a bit, so what you need to do actually is find the pipe.
Find the "right" side and the "left" side of the pipe, see which one
reaches the wall, and then only things on the right side of a pipe will
ever be inside. that is, a given pipe has a given direction and that
should be working